Removing PKG_CONFIG_PATH from .bash_profile
It seems to be causing issues with ruby/eventmachine compilation. We
were seeing the following when running bosh integration tests:

    /Users/cpi/.rvm/gems/ruby-2.4.2/gems/eventmachine-1.2.7/lib/em/connection.rb:459: [BUG] Segmentation fault at 0x0000000000019000

    -- Control frame information -----------------------------------------------
    c:0052 p:---- s:0284 e:000283 CFUNC  :start_tls
    c:0051 p:0287 s:0279 e:000278 METHOD /Users/cpi/.rvm/gems/ruby-2.4.2/gems/eventmachine-1.2.7/lib/em/connection.rb:459

After removing, we ran the following to rebuild our ruby version and
native gems without the openssl-dependency:

    rvm reinstall ruby-2.4.4
    bundle install --local --force

We theorize that ruby was being linked to openssl, but eventmachine/gems
may have ignored PKG_CONFIG_PATH and relied on system libssl libraries
which caused the segfaults when both tried to execute from the same
runtime and process space.

[#155593275](https://www.pivotaltracker.com/story/show/155593275)

Co-authored-by: Florian Nachtigall <florian.nachtigall@sap.com>